+++ R/geom-flow.r
@@ -28,7 +28,7 @@ geom_flow <- function(mapping = NULL,
                       position = "identity",
                       width = 1/3,
                       knot.pos = 1/4, knot.prop = TRUE,
-                      curve_type = "xspline", curve_range = NULL,
+                      curve_type = NULL, curve_range = NULL,
                       segments = NULL,
                       aes.flow = "forward",
                       na.rm = FALSE,
@@ -91,9 +91,14 @@ GeomFlow <- ggproto(
   draw_panel = function(self, data, panel_params, coord,
                         width = 1/3, aes.flow = "forward",
                         knot.pos = 1/4, knot.prop = TRUE,
-                        curve_type = "xspline", curve_range = NULL,
+                        curve_type = NULL, curve_range = NULL,
                         segments = NULL) {
-
+    
+    # parameter defaults
+    if (is.null(curve_type)) curve_type <- ggalluvial_opt("curve_type")
+    if (is.null(curve_range)) curve_range <- ggalluvial_opt("curve_range")
+    if (is.null(segments)) segments <- ggalluvial_opt("segments")
+    
     # exclude one-sided flows
     data <- data[complete.cases(data), ]