+++ R/stat-flow.r
@@ -34,12 +34,12 @@ stat_flow <- function(mapping = NULL,
                       data = NULL,
                       geom = "flow",
                       position = "identity",
-                      decreasing = ggalluvial_opt("decreasing"),
-                      reverse = ggalluvial_opt("reverse"),
-                      absolute = ggalluvial_opt("absolute"),
+                      decreasing = NULL,
+                      reverse = NULL,
+                      absolute = NULL,
                       discern = FALSE,
                       negate.strata = NULL,
-                      aes.bind = ggalluvial_opt("aes.bind"),
+                      aes.bind = NULL,
                       infer.label = FALSE,
                       min.y = NULL, max.y = NULL,
                       na.rm = FALSE,
@@ -134,14 +134,20 @@ StatFlow <- ggproto(
   },
   
   compute_panel = function(self, data, scales,
-                           decreasing = ggalluvial_opt("decreasing"),
-                           reverse = ggalluvial_opt("reverse"),
-                           absolute = ggalluvial_opt("absolute"),
+                           decreasing = NULL,
+                           reverse = NULL,
+                           absolute = NULL,
                            discern = FALSE, distill = first,
                            negate.strata = NULL,
-                           aes.bind = ggalluvial_opt("aes.bind"),
+                           aes.bind = NULL,
                            infer.label = FALSE,
                            min.y = NULL, max.y = NULL) {
+    
+    # parameter defaults
+    if (is.null(decreasing)) decreasing <- ggalluvial_opt("decreasing")
+    if (is.null(reverse)) reverse <- ggalluvial_opt("reverse")
+    if (is.null(absolute)) absolute <- ggalluvial_opt("absolute")
+    if (is.null(aes.bind)) aes.bind <- ggalluvial_opt("aes.bind")
     
     # introduce label
     if (infer.label) {